requirements.txt
aniso8601==9.0.1
appdirs==1.4.4
beautifulsoup4==4.11.1
bert-extractive-summarizer==0.10.1
blis==0.7.7
bs4==0.0.1
catalogue==2.0.7
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
cssselect==1.1.0
cymem==2.0.6
fake-useragent==0.1.11
filelock==3.6.0
Flask==2.1.1
Flask-RESTful==0.3.9
gensim==3.8.3
huggingface-hub==0.5.1
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
joblib==1.1.0
langcodes==3.3.0
lexrank==0.1.0
lxml==4.8.0
MarkupSafe==2.1.1
murmurhash==1.0.6
networkx==2.8
nltk==3.7
numpy==1.22.3
packaging==21.3
parse==1.19.0
path==16.4.0
path.py==12.5.0
pathy==0.6.1
platformdirs==2.5.2
preshed==3.0.6
pydantic==1.8.2
pyee==8.2.2
pyparsing==3.0.8
pyppeteer==1.0.2
pyquery==1.4.3
pyrsistent==0.18.1
pytz==2022.1
PyYAML==6.0
readability-lxml==0.8.1
regex==2022.3.15
requests==2.27.1
requests-html==0.10.0
sacremoses==0.0.49
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
smart-open==5.2.1
soupsieve==2.3.2
spacy==3.2.4
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.3
thinc==8.0.15
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.11.0
tqdm==4.64.0
transformers==4.18.0
typer==0.4.1
typing_extensions==4.2.0
uritools==4.0.0
urlextract==1.5.0
urllib3==1.26.9
w3lib==1.22.0
wasabi==0.9.1
websockets==10.3
Werkzeug==2.1.1
zipp==3.8.0
